requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Werkzeug==3.0.1
gunicorn==21.2.0
psycopg2-binary==2.9.10
python-dotenv==1.0.0
selenium==4.16.0
pytest==7.4.3
webdriver-manager==4.0.1
